Move the helper functions from The Nock
This helps when other files need to use these definitions
